"centros" array removed. "dibujarpuntos" edited.
"centros" array was a left-over from a previous stage when every repulsion force was focused just in the midpoint of each line. The actual system now discretizes the line in "discrete" segments and solves the interaction of each of these pieces of line with every other piece of every other line in the board.
The "dibujarpuntos" subroutine was altered so that a for every line the drawing number of steps is iqual to the line's length in pixels.